+++ 3 주차/1260번.py
