+++ 4 주차/11047번.py
