+++ 4 주차/2875번.py
