+++ 9 주차 카드구매하기&치킨배달/김서영/11052 번.py
