+++ 9 주차 카드구매하기&치킨배달/서동혁/15686.py
